Some intro_panels cleanup, move WellList to src/components
