+++ client/src/components/index.js
@@ -28,6 +28,7 @@ export { CountdownCircle } from './CountdownCircle.js';
 export { Countdown } from './Countdown.js';
 export { LogInteractionEvents } from './LogInteractionEvents.js';
 export { GlossaryIcon } from './GlossaryIcon.js';
+export { WellList } from './WellList.js';
 
 export { StatelessModal, FixedPopover } from './modals_and_popovers';
